scraper/requirements.txt
scrapling[fetchers]
supabase
python-dotenv
